Add logic to handle targetted scan
